packets: implement new packet changes with EPIC_9_6_5

+++ librzu/src/packets/AuthClient/Flat/TS_CA_SELECT_SERVER.h
@@ -5,7 +5,7 @@
 #pragma pack(push, 1)
 struct TS_CA_SELECT_SERVER : public TS_MESSAGE
 {
-	uint16_t server_idx;
+	uint32_t server_idx;
 	static const uint16_t packetID = 10023;
 };
 #pragma pack(pop)
